Update patch set 2
Patch Set 2:

> Why don't we just put such object on the Runtime object? Or the
 > compiler? It's not like we need to allocate tons of them...

Note: most of the handling code was already sanitized, in that the caller was storing things in a unique_ptr. Calin made a change recently and didn't do it though, and caused a leak. This is only enforcing the unique_ptr usage throughout.

The runtime actually doesn't care about the features (which is potentially a mistake, as we should reject oat files in some cases). So leaving it this way was the easiest and smallest change...

Patch-set: 2
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0