Update patch set 2
Patch Set 2:

Hi, I have mostly implementation design concerns. 

First of all to me scheduling should be applied on machine instructions not HInstruction. I guess this implementation is based on HInstruction to save compiler memory footprint and probably to reduce the size of the change. If there are other reason I would appreciate if you could list them.
Even if scheduling is done on HInstruction I think that scheduling highly depends on Register allocation. The same HInstruction can cost differently if use is on register on in memory for example. So I guess that this pass should be moved after register allocation or you want to use this pass to help register allocation by re-ordering instructions and give some more free registers to RA? I think it will be difficult to do. If it is moved after register allocation we may build dependency graph basing on generated location to provide better scheduling.
Then to me scheduling is highly depends on the way the HInstruction will be lowered. In other words any change in CodeGenerator potentially may affect the cost of instruction in scheduler. I like that LocationBuilder and InstructionCodeGenerator are located near to each other in code generator. I think that cost estimator should be also located in code generator. That way we make the life of developers easier to let them keep in mind that after changes in code generation he should also pay attention to possible cost change.

One more think, you wrote about performance and compile time impact, what is about compiler memory footprint? Is it changed a lot or not?

Hope it helps.

Patch-set: 2
Label: Verified=0