Update patch set 1
Patch Set 1:

Hi Calin, I guessed verify-at-runtime is a class flag not a method, right? Does it make sense to re-verify the whole class? Actually I wonder what is the reason to avoid adding new flag to oat file? oat-file size? In general we can use this flag to pass any information from one compilation to another if it is needed in the future or profile is expected for this purpose?

Patch-set: 1