Update patch set 1
Patch Set 1:

> Hi Calin, I guessed verify-at-runtime is a class flag not a method,
 > right? Does it make sense to re-verify the whole class? Actually I
 > wonder what is the reason to avoid adding new flag to oat file?
 > oat-file size? In general we can use this flag to pass any
 > information from one compilation to another if it is needed in the
 > future or profile is expected for this purpose?

Yes, it's a class flag. It's consistent with the approach we had so far. We only record verification info per class not method. Oat file size is a concern and I don't think it worth the extra space to record the info on a per method level. This situation should be the exception and not the norm.

Patch-set: 1
Reviewer: Calin Juravle <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0