Update patch set 1
Patch Set 1:

Guys, marking it as non-verified is also ok to me, I considered that case. The thing I though about was to reduce the launch tie to avoid re-verify the class we have already verified. All classes for example throwing not supported exception will go this case. Anyway, Calin, please let me know when you upload the patch and I will abandon this one in favor of that patch.

Patch-set: 1