Update patch set 1
Patch Set 1:

ok, will wait for the patch from Calin. What I'm afraid is that if we mark class as verify-at-runtime it is possible that no one method from this class will be compiled in AOT mode (in JIT it will compile). It also includes re-compilation of the methods in "second day usage" mode when AOT is compiling with profile. Is it expected or you plan to handle it in some way?
I really want to see the patch :) Sorry for bothering you guys.

Patch-set: 1