Update patch set 1
Patch Set 1:

> From a CL point, can you please fix the tests in a separate CL (as
 > that's in principal separate from the disassembler)?

Do you mean all changes to the tests (and those in this patch in particular), or just the improvements that Vladimir suggested? I'd rather keep this patch self-contained, i.e. you apply it, and all tests pass (even though some of them are, strictly speaking, broken).

 > I guess I'm looking at an old version, DDI0487A_a_armv8_arm_errata.pdf
 > . Looking at DDI0487A_d_armv8_arm.pdf, the distinction between MOV
 > and MOVW is based on whether the imm16 can or cannot be represented
 > in T1 or T2. I personally find that confusing. Does the "W" in MOVW
 > now mean "the Wrong encoding"?

I think the idea is that if you disassemble something, then assemble the output, and then continue the process, you always end up with the same data. If the immediate operand can be represented in T1 and/or T2, then there is an ambiguity, and you can potentially end up with different output than the data you have started with. On the other hand, if the immediate is not representable in both T1 and T2, then there is no ambiguity, hence the suffix is redundant, so the ARM ARM drops it, which makes the disassembly a little bit more readable.

Anyway, I suggest getting the latest ARM ARM vesion, which is issue A.j as of this writing. The file name is "DDI0487A_j_armv8_arm.pdf". And yes, as mentioned, the VIXL disassembler is strongly ARMv8-oriented.

Patch-set: 1