Update patch set 10
Patch Set 10:

> Lots of too long lines...

In addition, please follow other Google style guidelines as well (comments are like sentences, start with cap, end with period, no tabs, etc).

Also, is there a design doc you can link to? Seems we are adding quite some instructions along the common path, is that a done decision already, or do we want to toggle this on/off somehow (you refer to a string.h constant, but that is not very clear).

Lastly, seems you can split the CL in a part that introduces new instructions (utils, and don't forget the tests), and the one that does the actual compression change.

Patch-set: 10
Reviewer: Aart Bik <1074526@85c56323-6fa9-3386-8a01-6480fb634889>