Update patch set 11
Patch Set 11:

Aart, we currently don't have a design doc for this project, and I agree that we are adding a lot of instructions. However, what's added is an additional path to treat the compressed case vs the existing uncompressed case, and the number of executed instructions should be similar.

String.compareTo is special in that we can't just punt if one string is compressed and the other is not compressed since it has to return the difference between the strings. I've talked to Jessica and we can add a call out to a helper function instead of handling it all in assembly.

The toggle for turning this feature on/off exists to allow her to commit her changes before they were fully finished. It will be removed once it's finished.

Patch-set: 11
Reviewer: Jeff Hao <1010108@85c56323-6fa9-3386-8a01-6480fb634889>